refactor: crypto.randomUUID over the uuid dep, and one MSVC bootstrap
uuid had five call sites across two files, all v4. crypto.randomUUID() is
already used ten times elsewhere in the same tree — renderer runs in a
secure context, main runs Node 22. Drops the dependency.

findVcVarsAll() was duplicated byte-for-byte (modulo comments) between
build-windows-wgc-helper.mjs and build-windows-compositor-addon.mjs, as
was the promisified spawn. Both move to scripts/msvcEnv.mjs. The two
runInVsEnv bodies genuinely differ (SDK lib compat shims vs cargo) and
stay put. The two scripts drop from 197/187 lines to 99/93.

Kept deliberately: `motion`. Its spring() is a real integrator
(keyframes, velocity, stiffness, damping, restDelta, generator protocol)
driving cursor smoothing; hand-rolling one to shed a dependency is
writing code, not deleting it.

tsc clean; 94 files / 1108 tests pass.
