+++ package.json
@@ -101,7 +101,6 @@
 		"sonner": "^2.0.7",
 		"tailwind-merge": "^3.5.0",
 		"tailwindcss-animate": "^1.0.7",
-		"uuid": "^13.0.0",
 		"web-demuxer": "^4.0.0",
 		"zod": "^4.1.12",
 		"zustand": "^5.0.8"
