+++ scripts/build-windows-compositor-addon.mjs
@@ -15,125 +15,17 @@
 // shipped, and require() fails at runtime even with the right dir on PATH.
 // See poc-d3d/.cargo/config.toml for the pin.
 
-import { spawn, spawnSync } from "node:child_process";
 import fs from "node:fs";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
+import { findVcVarsAll, run as spawnStep } from "./msvcEnv.mjs";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const ROOT = path.join(__dirname, "..");
 const POC_D3D_DIR = path.join(ROOT, "poc-d3d");
 const BUILD_OUT_DIR = path.join(ROOT, "electron", "native", "compositor-view", "build");
 
-function findVcVarsAll() {
-	const explicit = process.env.VCVARSALL;
-	if (explicit && fs.existsSync(explicit)) {
-		return explicit;
-	}
-
-	const vswhere = "C:\\Program Files (x86)\\Microsoft Visual Studio\\Installer\\vswhere.exe";
-	if (fs.existsSync(vswhere)) {
-		const result = spawnSync(
-			vswhere,
-			[
-				"-latest",
-				"-products",
-				"*",
-				"-requires",
-				"Microsoft.VisualStudio.Component.VC.Tools.x86.x64",
-				"-property",
-				"installationPath",
-			],
-			{ encoding: "utf8", windowsHide: true },
-		);
-		const installPath = result.stdout?.trim();
-		if (result.status === 0 && installPath) {
-			const candidate = path.join(installPath, "VC", "Auxiliary", "Build", "vcvarsall.bat");
-			if (fs.existsSync(candidate)) {
-				return candidate;
-			}
-		}
-	}
-
-	if (process.env.VSINSTALLDIR) {
-		const candidate = path.join(
-			process.env.VSINSTALLDIR,
-			"VC",
-			"Auxiliary",
-			"Build",
-			"vcvarsall.bat",
-		);
-		if (fs.existsSync(candidate)) {
-			return candidate;
-		}
-	}
-
-	// vswhere doesn't always enumerate pre-release channels (e.g. "Insiders"
-	// builds) — walk the install roots generically so new VS releases and
-	// preview channels are found automatically (same policy as the wgc-capture
-	// helper build script).
-	const editions = ["Community", "Professional", "Enterprise", "BuildTools"];
-	const installRoots = [
-		"C:\\Program Files\\Microsoft Visual Studio",
-		"C:\\Program Files (x86)\\Microsoft Visual Studio",
-	];
-
-	const listDirs = (dir) => {
-		try {
-			return fs
-				.readdirSync(dir, { withFileTypes: true })
-				.filter((entry) => entry.isDirectory())
-				.map((entry) => path.join(dir, entry.name));
-		} catch {
-			return [];
-		}
-	};
-
-	for (const installRoot of installRoots) {
-		for (const versionDir of listDirs(installRoot)) {
-			for (const channelDir of [versionDir, ...listDirs(versionDir)]) {
-				const direct = path.join(channelDir, "VC", "Auxiliary", "Build", "vcvarsall.bat");
-				if (fs.existsSync(direct)) {
-					return direct;
-				}
-				for (const edition of editions) {
-					const nested = path.join(
-						channelDir,
-						edition,
-						"VC",
-						"Auxiliary",
-						"Build",
-						"vcvarsall.bat",
-					);
-					if (fs.existsSync(nested)) {
-						return nested;
-					}
-				}
-			}
-		}
-	}
-
-	return null;
-}
-
-function run(command, args, options = {}) {
-	return new Promise((resolve, reject) => {
-		const child = spawn(command, args, {
-			cwd: POC_D3D_DIR,
-			stdio: "inherit",
-			windowsHide: true,
-			...options,
-		});
-		child.once("error", reject);
-		child.once("exit", (code) => {
-			if (code === 0) {
-				resolve();
-			} else {
-				reject(new Error(`${command} ${args.join(" ")} failed with code ${code}`));
-			}
-		});
-	});
-}
+const run = (command, args, options = {}) => spawnStep(command, args, { cwd: ROOT, ...options });
 
 async function runInVsEnv(command) {
 	const vcvarsAll = findVcVarsAll();
