+++ scripts/build-windows-wgc-helper.mjs
@@ -1,8 +1,8 @@
-import { spawn, spawnSync } from "node:child_process";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
+import { findVcVarsAll, run as spawnStep } from "./msvcEnv.mjs";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const ROOT = path.join(__dirname, "..");
@@ -11,101 +11,6 @@ const BUILD_DIR = path.join(SOURCE_DIR, "build");
 const COMPAT_LIB_DIR = path.join(BUILD_DIR, "compat-libs");
 const BIN_DIR = path.join(ROOT, "electron", "native", "bin", "win32-x64");
 const CMAKE = process.env.CMAKE_EXE ?? "cmake";
-
-function findVcVarsAll() {
-	const explicit = process.env.VCVARSALL;
-	if (explicit && fs.existsSync(explicit)) {
-		return explicit;
-	}
-
-	const vswhere = "C:\\Program Files (x86)\\Microsoft Visual Studio\\Installer\\vswhere.exe";
-	if (fs.existsSync(vswhere)) {
-		const result = spawnSync(
-			vswhere,
-			[
-				"-latest",
-				"-products",
-				"*",
-				"-requires",
-				"Microsoft.VisualStudio.Component.VC.Tools.x86.x64",
-				"-property",
-				"installationPath",
-			],
-			{ encoding: "utf8", windowsHide: true },
-		);
-		const installPath = result.stdout?.trim();
-		if (result.status === 0 && installPath) {
-			const candidate = path.join(installPath, "VC", "Auxiliary", "Build", "vcvarsall.bat");
-			if (fs.existsSync(candidate)) {
-				return candidate;
-			}
-		}
-	}
-
-	if (process.env.VSINSTALLDIR) {
-		const candidate = path.join(
-			process.env.VSINSTALLDIR,
-			"VC",
-			"Auxiliary",
-			"Build",
-			"vcvarsall.bat",
-		);
-		if (fs.existsSync(candidate)) {
-			return candidate;
-		}
-	}
-
-	// vswhere doesn't always enumerate pre-release channels (e.g. "Insiders"
-	// builds), and the on-disk layout for those isn't a stable "<year>\<edition>"
-	// path -- Visual Studio 2026 Insiders installs under a numeric product
-	// version folder like "18\Insiders\<edition>" instead of "2026\<edition>".
-	// Walk the install roots generically instead of hard-coding version/channel
-	// names so new VS releases and preview channels are found automatically.
-	const editions = ["Community", "Professional", "Enterprise", "BuildTools"];
-	const installRoots = [
-		"C:\\Program Files\\Microsoft Visual Studio",
-		"C:\\Program Files (x86)\\Microsoft Visual Studio",
-	];
-
-	const listDirs = (dir) => {
-		try {
-			return fs
-				.readdirSync(dir, { withFileTypes: true })
-				.filter((entry) => entry.isDirectory())
-				.map((entry) => path.join(dir, entry.name));
-		} catch {
-			return [];
-		}
-	};
-
-	for (const installRoot of installRoots) {
-		for (const versionDir of listDirs(installRoot)) {
-			// versionDir is either an edition directly ("2022\Community") or a
-			// channel that nests editions ("18\Insiders\Community").
-			for (const channelDir of [versionDir, ...listDirs(versionDir)]) {
-				const direct = path.join(channelDir, "VC", "Auxiliary", "Build", "vcvarsall.bat");
-				if (fs.existsSync(direct)) {
-					return direct;
-				}
-				for (const edition of editions) {
-					const nested = path.join(
-						channelDir,
-						edition,
-						"VC",
-						"Auxiliary",
-						"Build",
-						"vcvarsall.bat",
-					);
-					if (fs.existsSync(nested)) {
-						return nested;
-					}
-				}
-			}
-		}
-	}
-
-	return null;
-}
 
 function findWindowsSdkUmLibDir() {
 	const sdkLibRoot = "C:\\Program Files (x86)\\Windows Kits\\10\\Lib";
@@ -122,24 +27,7 @@ function findWindowsSdkUmLibDir() {
 		.at(-1);
 }
 
-function run(command, args, options = {}) {
-	return new Promise((resolve, reject) => {
-		const child = spawn(command, args, {
-			cwd: ROOT,
-			stdio: "inherit",
-			windowsHide: true,
-			...options,
-		});
-		child.once("error", reject);
-		child.once("exit", (code) => {
-			if (code === 0) {
-				resolve();
-			} else {
-				reject(new Error(`${command} ${args.join(" ")} failed with code ${code}`));
-			}
-		});
-	});
-}
+const run = (command, args, options = {}) => spawnStep(command, args, { cwd: ROOT, ...options });
 
 async function runInVsEnv(command) {
 	const vcvarsAll = findVcVarsAll();
